--- integrations/github/clients/constants.py
@@ -1,1 +1,0 @@
-DEFAULT_PAGE_SIZE = 100
