+++ integrations/github/main.py
@@ -1,19 +1,16 @@
-from enum import StrEnum
 from typing import Any
 
-from clients.github import IntegrationClient
+from integration.clients.github import IntegrationClient
+from integration.utils.kind import ObjectKind
 from port_ocean.context.ocean import ocean
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
 
 
-class ObjectKind(StrEnum):
-    REPOSITORY = "repository"
-    PULL_REQUEST = "pull-request"
-    ISSUE = "issue"
-    TEAM = "team"
-    WORKFLOW = "workflow"
+def init_client() -> IntegrationClient:
+    return IntegrationClient()
 
 
+# resync all object kinds
 @ocean.on_resync()
 async def on_resync(kind: str) -> list[dict[Any, Any]]:
     if kind == ObjectKind.REPOSITORY:
@@ -22,9 +19,10 @@ async def on_resync(kind: str) -> list[dict[Any, Any]]:
     return []
 
 
+# resync repository
 @ocean.on_resync(ObjectKind.REPOSITORY)
 async def resync_repositories(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     if kind == ObjectKind.REPOSITORY:
-        client = IntegrationClient()
+        client = init_client()
         async for repository in client.get_repositories():
             yield repository
